fix(skin): split support_bundle tests under the 1000-line budget
Reset retry + skin dirs pushed support_bundle.rs over 1000. Tests move
to support_bundle_tests.rs so FILES_OVER_1K stays at 80.
